streamlit/requirements.txt
pandas
tweepy==3.10
pymongo
streamlit>=1.9.2 
python-decouple
matplotlib
scipy
dnspython
kafka-python